requirements.txt
aniso8601==9.0.1
blinker==1.6.2
boto3==1.28.52
botocore==1.31.52
click==8.1.7
cmdstanpy==1.1.0
contourpy==1.1.1
convertdate==2.4.0
cycler==0.11.0
ephem==4.1.4
flask==2.3.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
fonttools==4.42.1
gunicorn==21.2.0
holidays==0.33
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.5
LunarCalendar==0.0.9
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
prophet==1.1.4
PyMeeus==0.5.12
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.6.2
six==1.16.0
tenacity==8.2.3
tqdm==4.66.1
tzdata==2023.3
urllib3==1.26.16
werkzeug==2.3.7
zipp==3.17.0
